% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edaFile.R
\name{edatest}
\alias{edatest}
\title{edatest}
\usage{
edatest(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
average, median, stddev
}
\description{
This function calculates mean, median and Standard deviation of a numeric vector.
}
\examples{
edatest(c(seq(1:10)))
}
